verbs: Move all dynamic driver opening code to dynamic_driver.c
This is preparation for removing this code when statically linking.
